Dev branch

+++ src/character/views.py
@@ -2,11 +2,100 @@ from flask import render_template, Blueprint, request, redirect, url_for, jsonif
 from flask_login import current_user, login_required
 from sqlalchemy import select
 from src.auth.models import User
-from src.character.models import Character, Character_Class, DND_Class, DND_Race, DND_Background, Character_Details, Character_Stats, Character_Race, Character_Hit_Points, Character_Death_Saves
+from src.character.models import Character, Character_Class, DND_Class, DND_Race, DND_Background, Character_Details, Character_Stats, Character_Race, Character_Hit_Points, Character_Death_Saves, DND_Skill
 from src import db
 import math, json
 
 character_bp = Blueprint('character_bp', __name__, template_folder='../templates')
+
+def get_character_info(char_id) -> dict:
+    char_info = {}
+
+    #Load Base Character Info
+    char = Character.query.filter_by(char_id=char_id).first().__dict__
+    char.pop('_sa_instance_state', None)
+    char_info.update(char)
+
+    #Load Character Race and Race Based Character Stats
+    race = (db.session.execute(
+            select(Character_Race.race_id, DND_Race.name, DND_Race.speed, DND_Race.size)
+            .join(DND_Race, DND_Race.race_id == Character_Race.race_id)
+            .where(Character_Race.char_id == char_id)
+        ).first())
+    char_info.update({'race_id':race[0], 'race_name':race[1], 'speed':race[2], 'size':race[3]})
+
+    #Load Character Class(es)
+    classes = (db.session.execute(
+            select(Character_Class.class_id, Character_Class.class_level, DND_Class.name)
+            .join(DND_Class, DND_Class.class_id == Character_Class.class_id)
+            .where(Character_Class.char_id == char_id)
+            .order_by(Character_Class.class_id)
+        ))
+    char_info.update({'classes': []})
+    for cls in classes:
+        char_info['classes'].append({'class_id':cls[0], 'level':cls[1], 'class_name':cls[2]})
+
+    #Load Character Modifier Scores
+    mods = (db.session.execute(
+            select(Character_Stats)
+            .where(Character_Race.char_id == char_id)
+        ).first())
+    mods = Character_Stats.query.filter_by(char_id=char_id).first().__dict__
+    char_info.update({'modifier_scores': []})
+    mods.pop('_sa_instance_state', None); mods.pop('char_id', None)
+    for attribute in mods:
+        char_info['modifier_scores'].append({'modifier_name': attribute, 'score': mods[attribute], 'value': math.floor((mods[attribute] - 10) / 2)})
+
+    #Load Character Health
+    char_hp = Character_Hit_Points.query.filter_by(char_id=char_id).first().__dict__
+    char_hp.pop('_sa_instance_state', None)
+    char_info.update(char_hp)
+
+    #Load Character Death Saves
+    char_death_saves = Character_Death_Saves.query.filter_by(char_id=char_id).first().__dict__
+    char_death_saves.pop('_sa_instance_state', None)
+    char_info.update(char_death_saves)
+
+    #Load Character Skills
+    skills = (db.session.execute(
+            select(DND_Skill.skill_id, DND_Skill.skill_name, DND_Skill.modifier_type, DND_Skill.linked_proficiency_id)
+            .where(DND_Skill.is_offical == True)
+            .order_by(DND_Skill.skill_name, DND_Skill.modifier_type, DND_Skill.skill_id)
+        ))
+    char_info.update({'skills': []})
+    for skill in skills:
+        mod_type_conversion = ['strength', 'dexterity', 'constitution', 'intelligence', 'wisdom', 'charisma']
+        mod_value = 0
+        for mod in char_info['modifier_scores']:
+            if mod['modifier_name'] != mod_type_conversion[skill[2]]: continue
+            mod_value = mod['value']
+        char_info['skills'].append({'skill_id':skill[0], 'skill_name':skill[1], 'modifier_type':mod_type_conversion[skill[2]], 'modifier_value': mod_value, 'linked_proficiency_id':skill[3]})
+
+    #Load Character Armor Class
+    for mod in char_info['modifier_scores']:
+        if mod['modifier_name'] != 'dexterity': continue
+        char_info.update({'armor_class': mod['value'] + 10})
+
+    #Load Character Initiative
+    for mod in char_info['modifier_scores']:
+        if mod['modifier_name'] != 'dexterity': continue
+        char_info.update({'initiative': mod['value']})
+
+    #Load Character Passive Perception
+    for mod in char_info['modifier_scores']:
+        if mod['modifier_name'] != 'wisdom': continue
+        char_info.update({'passive_perception': mod['value'] + 10})
+
+    #print(char_info)
+    return char_info
+
+
+
+
+#{'proficiency_bonus': 5, 'inspiration': None, 'char_id': 21, 'owner_id': 8, 'name': 'Karnation2', 'total_level': 16, 'race_id': 1, 'race_name': 'Dragonborn', 'speed': 30, 'size': 'Medium', 'classes': [{'class_id': 3, 'level': 10, 'class_name': 'Cleric'}, {'class_id': 5, 'level': 3, 'class_name': 'Fighter'}, {'class_id': 6, 'level': 3, 'class_name': 'Monk'}], 'modifier_scores': [{'dexterity': 8, 'constitution': 8, 'wisdom': 8, 'strength': 8, 'intelligence': 8, 'charisma': 8}], 'hit_points': 0, 'temp_hit_points': 0, 'max_hit_points': 0, 'success_throws': 0, 'fail_throws': 0}
+
+
+
 
 @character_bp.route("/character/<character_id>", methods=['GET', 'POST'])
 def character(character_id):
@@ -55,9 +144,46 @@ def character(character_id):
         db.session.commit()  # Save the changes
         return redirect(url_for('character_bp.character', character_id=character_id))
 
-    return render_template('character/character.html', user=user, character=char, current_class=current_class, current_level=current_level, all_classes=all_classes)
+    return render_template('character/character.html', user=user, character=char, current_class=current_class, current_level=current_level, all_classes=all_classes, char_si=get_character_info(char.char_id))
 
+def calculate_max_hp(character_id):
+    character = Character.query.get(character_id) # Get the character by ID
+    if not character:
+        raise ValueError(f"Character with ID {character_id} not found.")
 
+    constitution_score = int(Character_Stats.query.filter_by(char_id=character_id).first().constitution) # Constitution score
+    constitution_modifier = (constitution_score - 10) // 2 # Constitution modifier
+
+    character_classes = Character_Class.query.filter_by(char_id=character_id).all() # Get all classes for the character
+
+    if not character_classes:
+        raise ValueError("Character has no associated classes.")
+
+    first_class = Character_Class.query.filter_by(char_id=character_id,is_initial_class=True).first()
+    first_class_data = DND_Class.query.get(first_class.class_id)
+    
+    if not first_class_data:
+        raise ValueError(f"Class with ID {first_class.class_id} not found.")
+
+    hit_die = first_class_data.hit_die # Hit die for the first class
+    first_level_hp = hit_die # First level HP is the max roll of the hit die
+
+    additional_hp = 0
+    total_levels = 0
+
+    for char_class in character_classes: # 
+        class_data = DND_Class.query.get(char_class.class_id) # Get class data
+        if not class_data:
+            continue
+
+        levels = char_class.class_level - 1 if char_class == first_class else char_class.class_level # Subtract 1 for first class since we already calculated first level HP
+        additional_hp += levels * (class_data.hit_die // 2 + 1)  # Average roll per level
+        total_levels += levels # Total levels calculation
+
+    con_bonus_hp = (total_levels + 1) * constitution_modifier # Add con bonus for each level
+
+    max_hp = first_level_hp + additional_hp + con_bonus_hp
+    return max(max_hp, 1)  # Ensure HP is at least 1
 
 @character_bp.route("/character/create", methods=['GET', 'POST'])
 def create():
@@ -72,8 +198,8 @@ def create():
         classes = []
         for i in range(len(levels)):
             if levels[i] != "0":
-                classes.append({'level':i+1, 'class_id':f'{all_classes[i].class_id}'})
-
+                classes.append({'level':levels[i], 'class_id':f'{all_classes[i].class_id}'})
+        initial_class = request.form.get('selectFirstClass')
         char_name = request.form.get('charname')
         #ruleset = request.form.get('ruleset')
         #xp_method = request.form.get('xp_method')
@@ -153,18 +279,30 @@ def create():
         db.session.add(new_character_race)
 
         for new_class in classes:
-            new_character_class = Character_Class(
-                char_id = new_character.char_id,
-                class_id = new_class['class_id'],
-                class_level = new_class['class_id']
-            )
+
+            if initial_class == new_class['class_id']:
+                    new_character_class = Character_Class(
+                    char_id = new_character.char_id,
+                    class_id = new_class['class_id'],
+                    class_level = new_class['level'],
+                    is_initial_class = True
+                )
+            else:
+                new_character_class = Character_Class(
+                    char_id = new_character.char_id,
+                    class_id = new_class['class_id'],
+                    class_level = new_class['level'],
+                    is_initial_class = False
+                )
             db.session.add(new_character_class)
+        
+        db.session.commit() # Needed in order to calculate max HP
 
         #TODO: Save Characters Actual HP (not just temp values)
         new_character_hp = Character_Hit_Points(
             char_id = new_character.char_id,
-            hit_points = 0,
-            max_hit_points = 0,
+            hit_points = calculate_max_hp(new_character.char_id),
+            max_hit_points = calculate_max_hp(new_character.char_id),
             temp_hit_points = 0
         )
         db.session.add(new_character_hp)
@@ -184,21 +322,25 @@ def create():
     
     return render_template("character/character_creator.html", all_backgrounds=all_backgrounds, all_classes=all_classes, all_races=all_races)
 
-@character_bp.route('/delete_character/<character_id>', methods=['POST'])
+@character_bp.route('/delete_character/<character_id>', methods=['POST', 'GET'])
 @login_required
 def delete_character(character_id):
-    # Find the character by ID
-    character = Character.query.filter_by(char_id=character_id, owner_id=current_user.id).first()
+    if request.method == 'POST':
+        # Find the character by ID
+        character = Character.query.filter_by(char_id=character_id, owner_id=current_user.id).first()
 
-    # This technically worked. But it's ugly so need some kind of solution to make it less ugly
-    if not character:
-        return jsonify({"error": "Character not found or unauthorized"}), 404
-    try:
-        # Delete the character from the database
-        db.session.delete(character)
-        db.session.commit()
-        return jsonify({"message": "Character deleted successfully"}), 200
-    except Exception as e:
-        print(e)
-        db.session.rollback()
-        return jsonify({"error": "An error occurred while deleting the character"}), 500
+        # This technically worked. But it's ugly so need some kind of solution to make it less ugly
+        if not character:
+            return jsonify({"error": "Character not found or unauthorized"}), 404
+        try:
+            # Delete the character from the database
+            db.session.delete(character)
+            db.session.commit()
+            return redirect(url_for('core.home'))
+            #return jsonify({"message": "Character deleted successfully"}), 200
+        except Exception as e:
+            print(e)
+            db.session.rollback()
+            return jsonify({"error": "An error occurred while deleting the character"}), 500
+    if request.method == 'GET':
+        return redirect(url_for('core_bp.home'))
